--- content-link-api/content-link-api/src/main/java/com/devgary/contentlinkapi/components/ApiException.kt
@@ -1,3 +1,0 @@
-package com.devgary.contentlinkapi.components
-
-class ApiException(message: String?) : Exception(message)
