--- content-link-api/content-link-api/src/main/java/com/devgary/contentlinkapi/components/gfycat/api/model/GfycatAuthenticationRequest.kt
@@ -1,9 +1,0 @@
-package com.devgary.contentlinkapi.components.gfycat.api.model
-
-import com.squareup.moshi.Json
-
-data class GfycatAuthenticationRequest(
-    @Json(name = "client_id") private val clientId: String,
-    @Json(name = "client_secret") private val clientSecret: String,
-    @Json(name = "grant_type") private var grantType: String = "client_credentials",
-)
