--- content-link-api/content-link-api/src/main/java/com/devgary/contentlinkapi/components/gfycat/api/model/GfycatAuthenticationResponse.kt
@@ -1,11 +1,0 @@
-package com.devgary.contentlinkapi.components.gfycat.api.model
-
-import com.squareup.moshi.Json
-
-data class GfycatAuthenticationResponse(
-    @Json(name = "token_type") private val tokenType: String? = null,
-    @Json(name = "scope") private val scope: String? = null,
-    @Json(name = "expires_in") val expiresIn: Int,
-    @Json(name = "access_token") val accessToken: String,
-    var expiryTimeUnix: Long = 0
-)
