--- content-link-api/content-link-api/src/main/java/com/devgary/contentlinkapi/components/gfycat/api/model/GfycatItemResponse.kt
@@ -1,7 +1,0 @@
-package com.devgary.contentlinkapi.components.gfycat.api.model
-
-import com.squareup.moshi.Json
-
-data class GfycatItemResponse(
-    @Json(name = "gfyItem") var gfycatItem: GfycatItem,
-)
