--- content-link-api/content-link-api/src/main/java/com/devgary/contentlinkapi/components/imgur/api/model/ImgurAlbumResponse.kt
@@ -1,12 +1,0 @@
-package com.devgary.contentlinkapi.components.imgur.api.model
-
-import com.squareup.moshi.Json
-
-data class ImgurAlbumResponse(
-    @Json(name = "data")
-    val album: ImgurAlbum,
-    @Json(name = "status")
-    val status: Int,
-    @Json(name = "success")
-    val success: Boolean
-)
