--- content-link-api/content-link-api/src/main/java/com/devgary/contentlinkapi/components/interfaces/ClearableMemory.kt
@@ -1,5 +1,0 @@
-package com.devgary.contentlinkapi.components.interfaces
-
-interface ClearableMemory {
-    fun clearMemory()
-}
