--- content-link-api/content-link-api/src/main/java/com/devgary/contentlinkapi/components/streamable/StreamableLinkParser.kt
@@ -1,29 +1,0 @@
-package com.devgary.contentlinkapi.components.streamable
-
-import com.devgary.contentlinkapi.util.LinkUtils
-
-class StreamableLinkParser {
-    data class StreamableLinkInfo(val shortCode: String?)
-    
-    companion object {
-        fun parse(url: String): StreamableLinkInfo {
-            LinkUtils.parseAlphanumericIdFromUrl(
-                url = url, 
-                startFromOccurrence = "streamable.com/s/", 
-                minLength = 3
-            )?.let {
-                return StreamableLinkInfo(shortCode = it)
-            }
-
-            LinkUtils.parseAlphanumericIdFromUrl(
-                url = url,
-                startFromOccurrence = "streamable.com/",
-                minLength = 3
-            )?.let {
-                return StreamableLinkInfo(shortCode = it)
-            }
-
-            return StreamableLinkInfo(null)
-        }
-    }
-}
