--- content-link-api/content-link-api/src/main/java/com/devgary/contentlinkapi/components/streamable/api/model/StreamableVideoResponse.kt
@@ -1,13 +1,0 @@
-package com.devgary.contentlinkapi.components.streamable.api.model
-
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class StreamableVideoResponse(
-    var url: String,
-    @Json(name = "thumbnail_url")
-    var thumbnailUrl: String? = null,
-    var title: String? = null,
-    var videos: MutableList<StreamableVideo> = mutableListOf()
-)
