+++ content-link-api/content-link-api/src/main/java/com/devgary/contentlinkapi/content/CompositeContentLinkHandler.kt
@@ -1,5 +1,5 @@
 package com.devgary.contentlinkapi.content
 
-import com.devgary.contentlinkapi.components.interfaces.ClearableMemory
+import com.devgary.contentlinkapi.handlers.interfaces.ClearableMemory
 
 interface CompositeContentLinkHandler : ContentLinkHandler, ClearableMemory
