+++ content-link-api/content-link-api/src/main/java/com/devgary/contentlinkapi/handlers/ApiException.kt
@@ -1,0 +1,3 @@
+package com.devgary.contentlinkapi.handlers
+
+class ApiException(message: String?) : Exception(message)
