+++ content-link-api/content-link-api/src/main/java/com/devgary/contentlinkapi/handlers/gfycat/api/model/GfycatItemResponse.kt
@@ -1,0 +1,7 @@
+package com.devgary.contentlinkapi.handlers.gfycat.api.model
+
+import com.squareup.moshi.Json
+
+data class GfycatItemResponse(
+    @Json(name = "gfyItem") var gfycatItem: GfycatItem,
+)
