+++ content-link-api/content-link-api/src/main/java/com/devgary/contentlinkapi/handlers/interfaces/ClearableMemory.kt
@@ -1,0 +1,5 @@
+package com.devgary.contentlinkapi.handlers.interfaces
+
+interface ClearableMemory {
+    fun clearMemory()
+}
